requirements.txt
django
djangorestframework
django-environ
mysqlclient
pyjwt
pymongo
azure-storage-blob